Implement validation
